Moved stats config to config file
