+++ config.example.js
@@ -45,4 +45,9 @@ module.exports = {
      * otherwise return the error message to be displayed on the PR
      */
   },
+  metrics: {
+    host: process.env.NODE_ENV === 'production' ? 'platform-statsd' : 'localhost',
+    port: 8125,
+    prefix: 'myname_landkid', // This is optional
+  },
 };
