+++ src/index.ts
@@ -17,7 +17,7 @@ import { routes } from './routes';
 import { initializePassport } from './auth/bitbucket';
 import { LandRequestHistory } from './lib/History';
 import { Logger } from './lib/Logger';
-import { stats } from './lib/Stats';
+import { stats } from './lib/utils/stats';
 
 const RedisStore = connectRedis(session);
 
@@ -64,7 +64,7 @@ async function main() {
 
   await routes(server, client, runner);
 
-  stats.increment('startup');
+  stats.increment('.startup');
 
   // TODO: lookup all admins in user service to add them to the redis cache
