+++ src/lib/Runner.ts
@@ -15,7 +15,7 @@ import {
   BannerMessageState,
 } from '../db';
 import { permissionService } from './PermissionService';
-import { stats } from '../lib/Stats';
+import { stats } from './utils/stats';
 
 // const MAX_WAITING_TIME_FOR_PR_MS = 2 * 24 * 60 * 60 * 1000; // 2 days - max time build can "land-when able"
 
@@ -226,10 +226,10 @@ export class Runner {
         pullRequestId,
         lockId,
       });
-      stats.increment('pull_request.land.success');
+      stats.increment('.pull_request.merge.success');
       return landRequest.setStatus('success');
     } catch (err) {
-      stats.increment('pull_request.land.fail');
+      stats.increment('.pull_request.merge.fail');
       return landRequest.setStatus('fail', 'Unable to merge pull request');
     }
   };
