+++ src/routes/bitbucket/proxy/index.ts
@@ -6,7 +6,7 @@ import { config } from '../../../lib/Config';
 import { LandRequestOptions } from '../../../types';
 import { Runner } from '../../../lib/Runner';
 import { Logger } from '../../../lib/Logger';
-import { stats } from '../../../lib/Stats';
+import { stats } from '../../../lib/utils/stats';
 
 export function proxyRoutes(runner: Runner, client: BitbucketClient) {
   const router = express();
@@ -83,7 +83,7 @@ export function proxyRoutes(runner: Runner, client: BitbucketClient) {
       });
 
       if (!pullRequestId || !aaid || !commit) {
-        stats.increment('pull_request.queue.fail');
+        stats.increment('.pull_request.queue.fail');
 
         res.sendStatus(404);
         return;
@@ -107,7 +107,7 @@ export function proxyRoutes(runner: Runner, client: BitbucketClient) {
         landRequest,
       });
 
-      stats.increment('pull_request.queue.success');
+      stats.increment('.pull_request.queue.success');
 
       res.sendStatus(200);
       runner.next();
@@ -125,7 +125,7 @@ export function proxyRoutes(runner: Runner, client: BitbucketClient) {
       });
 
       if (!pullRequestId || !aaid || !commit) {
-        stats.increment('pull_request.queue_when_able.fail');
+        stats.increment('.pull_request.queue_when_able.fail');
 
         res.sendStatus(400);
         return;
@@ -148,7 +148,7 @@ export function proxyRoutes(runner: Runner, client: BitbucketClient) {
         landRequest,
       });
 
-      stats.increment('pull_request.queue_when_able.success');
+      stats.increment('.pull_request.queue_when_able.success');
 
       res.sendStatus(200);
     }),
